Rename package from avatar-otel to video-ai-telemetry
Renames the Python package from avatar_otel to video_ai_telemetry
across all source files, tests, pyproject.toml, README, and plan.
