+++ src/video_ai_telemetry/_registry.py
@@ -1,4 +1,4 @@
-"""Module-level singletons for the avatar-otel SDK.
+"""Module-level singletons for the video-ai-telemetry SDK.
 
 Holds references to the active config, providers, and components
 so that module-level functions (pipeline_stage, info, etc.) can
@@ -10,7 +10,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from avatar_otel.config import AvatarOtelConfig
+    from video_ai_telemetry.config import AvatarOtelConfig
 
 _config: AvatarOtelConfig | None = None
 _tracer = None
@@ -21,5 +21,5 @@ _sdk = None
 
 def get_config() -> AvatarOtelConfig:
     if _config is None:
-        raise RuntimeError("avatar_otel.init() has not been called yet")
+        raise RuntimeError("video_ai_telemetry.init() has not been called yet")
     return _config
