+++ tests/test_eventloop.py
@@ -6,7 +6,7 @@ import time
 
 import pytest
 
-from avatar_otel.instrumentation.eventloop import (
+from video_ai_telemetry.instrumentation.eventloop import (
     install_eventloop_monitor,
     uninstall_eventloop_monitor,
 )
@@ -116,7 +116,7 @@ def test_default_logger_warning(caplog):
     install_eventloop_monitor(threshold_ms=50.0, warning_callback=None)
 
     try:
-        with caplog.at_level(logging.WARNING, logger="avatar_otel.eventloop"):
+        with caplog.at_level(logging.WARNING, logger="video_ai_telemetry.eventloop"):
             loop = asyncio.new_event_loop()
 
             async def blocking_coro():
